Revert "Refactor load dna file"
This reverts commit 4a6a7661802c2254cfeeea6f577d43833899ec15.
